# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pacific BioArchive API
  version: 1.0.0
  description: Contract-first API shared by the AWS entry point and Azure data service.
servers:
  - url: http://localhost:8000
    description: Local AWS-side API
...
paths:
...
  /uploads/reservations:
    post:
      tags: [uploads]
      operationId: reserveUpload
      security: [{CognitoBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "./schemas/upload-reservation-request.schema.json"}
      responses:
        "200":
          description: Existing duplicate or new upload reservation.
          content:
            application/json:
              schema: {$ref: "./schemas/upload-reservation-response.schema.json"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /media/tags:
    post:
      tags: [management]
      operationId: updateMediaTags
      security: [{CognitoBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "./schemas/bulk-tag-operation.schema.json"}
      responses:
        "200":
          description: Per-media update outcomes.
          content:
            application/json:
              schema: {$ref: "./schemas/tag-update-response.schema.json"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /media:
    get:
      tags: [media]
      operationId: listMedia
      security: [{CognitoBearer: []}]
      responses:
        "200":
          description: Media belonging to the authenticated user.
          content:
            application/json:
              schema: {$ref: "./schemas/query-response.schema.json"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    delete:
      tags: [management]
      operationId: deleteMedia
      security: [{CognitoBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "./schemas/delete-request.schema.json"}
      responses:
        "200":
          description: Per-media deletion outcomes.
          content:
            application/json:
              schema: {$ref: "./schemas/delete-response.schema.json"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
components:
  securitySchemes:
    CognitoBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito access token.
  schemas:
...
    HttpValidationError:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            required: [loc, msg, type]
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - {type: string}
                    - {type: integer}
              msg: {type: string}
              type: {type: string}
              input: {}
  responses:
    Unauthorized:
      description: Missing or invalid Cognito access token.
      content:
        application/json:
          schema: {$ref: "./schemas/error-response.schema.json"}
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            oneOf:
              - {$ref: "./schemas/error-response.schema.json"}
              - {$ref: "#/components/schemas/HttpValidationError"}
...